refactor(CrossAppLogin): Extract most of IntroViewModel to CrossAppLoginViewModel
This is to then remove the "crossAppLogin" prefixes from
properties and functions, avoid mixing this feature with
other things, and also have it the same as in kDrive.
